server/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
websockets>=16.0
redis>=5.0
asyncpg>=0.29
pyjwt>=2.8
bcrypt>=4.1
pydantic>=2.5
python-dotenv>=1.0

# Development/Testing
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
